+++ api/package.json
@@ -7,12 +7,13 @@
     "build": "wrangler deploy --dry-run --outdir ./build"
   },
   "dependencies": {
+    "cheerio": "^1.0.0-rc.12",
     "hono": "^3.2.2",
+    "nanoid": "^4.0.2",
     "ofetch": "^1.0.1",
+    "parse-url": "^8.1.0",
     "unstorage": "^1.6.1",
     "webscout": "*",
-    "cheerio": "^1.0.0-rc.12",
-    "parse-url": "^8.1.0",
     "zod": "^3.21.4"
   },
   "devDependencies": {
